Sanitize map name in the replay browser.
