+++ common/maps.ts
@@ -197,6 +197,25 @@ export function getTeamNames(
   }
 }
 
+/**
+ * Filters out unprintable characters used for color codes in BW (we don't utilize these in our
+ * client, and they just show up as tofu).
+ */
+export function filterColorCodes(str: string) {
+  return Array.from(str)
+    .filter(c => {
+      const code = c.charCodeAt(0)
+      return (
+        code > 0x1f ||
+        /** newline */
+        code === 0x0a ||
+        /** carriage return */
+        code === 0x0d
+      )
+    })
+    .join('')
+}
+
 export interface UploadMapResponse {
   map: MapInfoJson
 }
